+++ CassoEmuCore/Machines/Apple2/Apple2/Apple2.h
@@ -38,6 +38,7 @@ public:
     int   GetSlotCount         () const override { return (7); }
     int   GetGamePortAxisCount () const override { return (4); }
     bool  HasGamePortDevice    () const override { return (true); }
+    bool  HasAnnunciators      () const override { return (true); }
     bool  HasCaseSwitches    () const override { return (false); }
     bool  HasBuiltInDrive    () const override { return (false); }
 };