+++ CassoEmuCore/Ui/Settings/SettingsApplyAdapter.h
@@ -46,6 +46,7 @@ public:
     void ApplyWriteProtect           (int drive, bool wp)                    override;
     void ApplyExternalDriveConnected (bool connected)                        override;
     void ApplyMouseConnected         (bool connected)                        override;
+    void ApplyGamePortAdapter        (GamePortAdapter adapter)               override;
     void QueueMachineReset           ()                                      override { m_resetQueued = true; }
 
     bool IsResetQueued () const { return m_resetQueued; }